% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscaLCR.R
\name{gscaLCR}
\alias{gscaLCR}
\title{The 2nd and 3rd step of gscaLCR, which are the partitioning and fitting regression.}
\usage{
gscaLCR(results.obj, covnames, multinomial.test.ref = "MAX")
}
\arguments{
\item{results.obj}{the results of gscaLCA.}

\item{covnames}{a vector of characters of covariate names.}

\item{multinomial.test.ref}{A character element. "MAX", "MIX", "FIRST", "LAST are the available options. The default is \code{MAX}.}
}
\value{
A results of class gscaLCA with covariates
}
\description{
The 2nd and 3rd step of gscaLCR, which are the partitioning and fitting regression.
}
\examples{
R2 = gscaLCA (dat = AddHealth[1:500, ],
               varnames = names(AddHealth)[2:6],
               ID.var = "AID",
               num.cluster = 3,
               num.factor = "EACH",
               Boot.num = 0,
               multiple.Core = F)

R2.gender = gscaLCR (R2, covnames = "Gender")
summary(R2.gender,  "multinomial.hard") # hard partitioning with multinomial regression
summary(R2.gender,  "multinomial.soft") # soft partitioning with multinomial regression
summary(R2.gender,  "binomial.hard")    # hard partitioning with binomial regression
summary(R2.gender,  "binomial.soft")    # soft partitioning with binomial regression

}
