% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gscaLCA.R
\name{summary.gscaLCA}
\alias{summary.gscaLCA}
\title{Summary of gscaLCA output or gscaLCR output}
\usage{
\method{summary}{gscaLCA}(object, print.cov.output = NULL, ...)
}
\arguments{
\item{object}{the object of gscaLCA or gscaLCR}

\item{print.cov.output}{a character of what type partitioning and regression. Four possible option are possible "multinomial.hard", "multinomial.soft", "binomial.hard", and "binomial.soft".}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
print model fit, prevalence, item probabilities, and regression results
}
\description{
Summary of gscaLCA output or gscaLCR output
}
\examples{
# summary(R2)
}
